Refactor FirestoreChatsClient
