+++ Owl/Models/Chat/GetMessages/GetNextMessagesResponse.swift
@@ -1,0 +1,14 @@
+//
+//  GetNextMessagesResponse.swift
+//  Owl
+//
+//  Created by Anastasia Holovash on 05.08.2022.
+//
+
+import Foundation
+import Firebase
+
+struct GetNextMessagesResponse: Equatable {
+    let messageResponse: [MessageResponse]
+    let lastDocumentSnapshot: DocumentSnapshot?
+}
