+++ Owl/Modules/Root/Clients/Firestore/FirestoreChatsClient.swift
@@ -9,7 +9,6 @@ import Combine
 import ComposableArchitecture
 import Firebase
 import FirebaseFirestoreCombineSwift
-import FirebaseAuth
 
 struct FirestoreChatsClient {
 
@@ -30,248 +29,4 @@ struct FirestoreChatsClient {
     var subscribeForNewMessages: (DocumentSnapshot) -> Effect<[MessageResponse], NSError>
     var getNextMessages: (DocumentSnapshot) -> Effect<GetNextMessagesResponse, NSError>
     var sendMessage: (NewMessage) -> Effect<Bool, NSError>
-}
-
-struct GetLastMessagesResponse: Equatable {
-    let messageResponse: [MessageResponse]
-    let lastDocumentSnapshot: DocumentSnapshot
-    let subscribeForNewMessagesSnapshot: DocumentSnapshot
-}
-
-struct GetNextMessagesResponse: Equatable {
-    let messageResponse: [MessageResponse]
-    let lastDocumentSnapshot: DocumentSnapshot?
-}
-
-// MARK: - Live
-
-extension FirestoreChatsClient {
-
-    // swiftlint:disable function_body_length
-    static func live(userClient: UserClient) -> Self {
-        let openedChatId = CurrentValueSubject<String?, Never>(nil)
-        return Self(
-            openedChatId: openedChatId,
-            getChats: {
-                Effect.run { subscriber in
-                    guard let authUser = userClient.authUser.value else {
-                        subscriber.send(completion: .failure(.init(domain: "No user", code: 1)))
-                        return Empty<Any, NSError>(completeImmediately: true)
-                            .sink()
-                    }
-                    return Collection.chats.whereField("members", arrayContains: authUser.uid)
-                        .snapshotPublisher()
-                        .on(
-                            value: { snapshot in
-                                subscriber.send(
-                                    snapshot.documents.compactMap { document in
-                                        do {
-                                            return try document.data(as: ChatsListPrivateItem.self)
-                                        } catch let error as NSError {
-                                            subscriber.send(completion: .failure(error))
-                                            return nil
-                                        }
-                                    }
-                                )
-                            },
-                            error: { subscriber.send(completion: .failure($0 as NSError)) }
-                        )
-                        .sink()
-                }
-            },
-            chatWithUser: { uid in
-                Effect.future { callback in
-                    guard let authUser = userClient.authUser.value else {
-                        return callback(.failure(.init(domain: "No user", code: 1)))
-                    }
-                    let users = [uid, authUser.uid]
-                    let usersReversed: [String] = users.reversed()
-                    Collection.chats
-                        .whereField("members", in: [users, usersReversed])
-                        .getDocuments()
-                        .on(
-                            value: { snapshot in
-                                if let document = snapshot.documents.first {
-                                    do {
-                                        let chatsListPrivateItem = try document.data(as: ChatsListPrivateItem.self)
-                                        callback(.success(.chatItem(chatsListPrivateItem)))
-                                    } catch let error as NSError {
-                                        callback(.failure(error as NSError))
-                                    }
-                                } else {
-                                    callback(.success(.needToCreate(withUserID: uid)))
-                                }
-                            },
-                            error: { callback(.failure($0 as NSError)) }
-                        )
-                        .sink()
-                        .store(in: &cancellables)
-                }
-            },
-            createPrivateChat: { privateChatRequest in
-                Effect.future { callback in
-                    let newDocument = Collection.chats.document()
-
-                    // Update id of PrivateChatRequest
-                    var privateChatRequest = privateChatRequest
-                    privateChatRequest.id = newDocument.documentID
-
-                    newDocument.setData(from: privateChatRequest)
-                        .catch { error -> AnyPublisher<Void, Never> in
-                            callback(.failure(error as NSError))
-                            return Empty(completeImmediately: true)
-                                .eraseToAnyPublisher()
-                        }
-                        .flatMap { _ in newDocument.getDocument().eraseToAnyPublisher() }
-                        .on(
-                            value: { response in
-                                do {
-                                    let chatsListPrivateItem = try response.data(as: ChatsListPrivateItem.self)
-                                    callback(.success(chatsListPrivateItem))
-                                } catch let error as NSError {
-                                    callback(.failure(error as NSError))
-                                }
-                            },
-                            error: { callback(.failure($0 as NSError)) }
-                        )
-                        .sink()
-                        .store(in: &cancellables)
-                }
-            },
-            getLastMessages: {
-                Effect.future { callback in
-                    guard let chatID = openedChatId.value else {
-                        return
-                    }
-
-                    Collection.chatsMessages.document(chatID).collection("messages")
-                        .order(by: "sentAt", descending: true)
-                        .limit(to: 25)
-                        .getDocuments()
-                        .on { snapshot in
-                            guard
-                                let lastDocumentSnapshot = snapshot.documents.last,
-                                let subscribeForNewMessagesSnapshot = snapshot.documents.first
-                            else {
-                                return
-                            }
-
-                            let items = snapshot.documents.compactMap { document -> MessageResponse? in
-                                do {
-                                    return try document.data(as: MessageResponse.self)
-                                } catch let error as NSError {
-                                    callback(.failure(error))
-                                    return nil
-                                }
-                            }
-                            let response = GetLastMessagesResponse(
-                                messageResponse: items,
-                                lastDocumentSnapshot: lastDocumentSnapshot,
-                                subscribeForNewMessagesSnapshot: subscribeForNewMessagesSnapshot
-                            )
-                            callback(.success(response))
-                        } error: { error in
-                            callback(.failure(error as NSError))
-                        }
-                        .sink()
-                        .store(in: &cancellables)
-                }
-            },
-            subscribeForNewMessages: { snapshot in
-                Effect.run { subscriber in
-                    guard let chatID = openedChatId.value else {
-                        return Empty<Any, NSError>(completeImmediately: true)
-                            .sink()
-                    }
-
-                    return Collection.chatsMessages.document(chatID).collection("messages")
-                        .order(by: "sentAt", descending: true)
-                        .end(beforeDocument: snapshot)
-                        .snapshotPublisher()
-                        .on { snapshot in
-                            let items = snapshot.documents.compactMap { document -> MessageResponse? in
-                                do {
-                                    return try document.data(as: MessageResponse.self)
-                                } catch let error as NSError {
-                                    subscriber.send(completion: .failure(error))
-                                    return nil
-                                }
-                            }
-                            subscriber.send(items)
-                        }
-                        error: { error in
-                            subscriber.send(completion: .failure(error as NSError))
-                        }
-                        .sink()
-                }
-            },
-            getNextMessages: { snapshot in
-                Effect.future { callback in
-                    guard let chatID = openedChatId.value else {
-                        return
-                    }
-
-                    Collection.chatsMessages.document(chatID).collection("messages")
-                        .order(by: "sentAt", descending: true)
-                        .start(afterDocument: snapshot)
-                        .limit(to: 25)
-                        .getDocuments()
-                        .on { snapshot in
-                            let items = snapshot.documents.compactMap { document -> MessageResponse? in
-                                do {
-                                    return try document.data(as: MessageResponse.self)
-                                } catch let error as NSError {
-                                    callback(.failure(error))
-                                    return nil
-                                }
-                            }
-                            let response = GetNextMessagesResponse(
-                                messageResponse: items,
-                                lastDocumentSnapshot: snapshot.documents.last
-                            )
-                            callback(.success(response))
-                        }
-                        error: { error in
-                            callback(.failure(error as NSError))
-                        }
-                        .sink()
-                        .store(in: &cancellables)
-                }
-            },
-            sendMessage: { newMessage in
-                Effect.future { callback in
-                    let batch = FirebaseClient.firestore.batch()
-
-                    let newDocument = Collection.chatsMessages.document(newMessage.chatId).collection("messages").document()
-                    var message = newMessage.message
-                    message.id = newDocument.documentID
-
-                    let chat = Collection.chats.document(newMessage.chatId)
-
-                    do {
-                        try batch.setData(from: message, forDocument: newDocument)
-                        let encodedMessage = try Firestore.Encoder().encode(message)
-                        batch.updateData(["lastMessage": encodedMessage], forDocument: chat)
-
-                        batch.commit()
-                            .on { _ in
-                                callback(.success(true))
-                            }
-                            error: { error in
-                                callback(.failure(error as NSError))
-                            }
-                            .sink()
-                            .store(in: &cancellables)
-
-                    } catch {
-                        callback(.failure(error as NSError))
-                    }
-                }
-            }
-        )
-    }
-
-    static func messages(query: Query) {
-
-    }
 }
